feat(onboarding): share setup controls with settings
