+++ ThruRNDIS/Views/Settings/WireGuardView.swift
@@ -26,37 +26,15 @@ struct WireGuardView: View {
                 }
             }
 
-            Section("Network Extension") {
-                LabeledContent("Status") {
-                    SettingsStatusLabel(
-                        title: wireGuardSession.systemExtensionStatus.title,
-                        appearance: systemExtensionStatusAppearance
-                    )
-                }
-
-                Text(systemExtensionStatusDetail)
+            Section {
+                NetworkExtensionPermissionView()
+            } header: {
+                Text("Network Extension")
+            } footer: {
+                Text("Network Extension permission is required for WireGuard connections.")
                     .font(.callout)
                     .foregroundStyle(.secondary)
                     .fixedSize(horizontal: false, vertical: true)
-
-                HStack {
-                    Button("Request Activation") {
-                        store.requestWireGuardSystemExtensionActivation()
-                    }
-                    .disabled(!store.canRequestWireGuardSystemExtensionActivation)
-
-                    Button("Open Settings") {
-                        store.openWireGuardSystemExtensionSettings()
-                    }
-                    .buttonStyle(.link)
-
-                    Spacer()
-
-                    Button("Refresh Status") {
-                        store.refreshWireGuardSystemExtensionStatus()
-                    }
-                    .disabled(wireGuardSession.systemExtensionStatus.isTransitioning)
-                }
             }
 
             Section("Connection") {
@@ -198,48 +176,4 @@ struct WireGuardView: View {
             .foregroundStyle(.red)
     }
 
-    private var systemExtensionStatusDetail: LocalizedStringKey {
-        if wireGuardSession.systemExtensionStatus == .uninstalling {
-            return "Restart macOS to finish removing the Network Extension before requesting activation again."
-        }
-        if !store.runtimeEntitlements.systemExtensionInstall {
-            return "This build cannot activate the Network Extension. Run a signed copy of ThruRNDIS from Applications."
-        }
-
-        return switch wireGuardSession.systemExtensionStatus {
-        case .unknown:
-            "The Network Extension status has not been checked yet."
-        case .checking:
-            "Checking whether the Network Extension is active."
-        case .inactive:
-            "Request activation, then allow ThruRNDIS in System Settings before connecting."
-        case .activationRequested, .awaitingUserApproval:
-            "Activation was requested. Approve the Network Extension in System Settings."
-        case .active:
-            "The Network Extension is active and ready to connect."
-        case .uninstalling:
-            "Restart macOS to finish removing the Network Extension before requesting activation again."
-        case .restartRequired:
-            "Restart macOS to finish activating the Network Extension."
-        case .failed:
-            "The Network Extension status could not be determined."
-        }
-    }
-
-    private var systemExtensionStatusAppearance: SettingsStatusAppearance {
-        switch wireGuardSession.systemExtensionStatus {
-        case .active:
-            .active
-        case .checking, .activationRequested:
-            .transitioning
-        case .inactive:
-            .inactive
-        case .awaitingUserApproval, .uninstalling, .restartRequired:
-            .attention
-        case .failed:
-            .failed
-        case .unknown:
-            .unknown
-        }
-    }
 }
